Desenvolvimento EMS 0007 Consulta Encalhe

+++ nds-client/src/main/java/br/com/abril/nds/repository/MovimentoEstoqueCotaRepository.java
@@ -15,7 +15,7 @@ import br.com.abril.nds.model.estoque.TipoMovimentoEstoque;
 
 public interface MovimentoEstoqueCotaRepository extends Repository<MovimentoEstoqueCota, Long> {
 
-	public BigDecimal obterQtdProdutoEdicaoEncalhe(FiltroConsultaEncalheDTO filtro, Long idTipoMovimento, boolean indQtdEncalheAposPrimeiroDia);
+	public Integer obterQtdProdutoEdicaoEncalhe(FiltroConsultaEncalheDTO filtro, Long idTipoMovimento, boolean indQtdEncalheAposPrimeiroDia);
 	
 	public BigDecimal obterQtdItemProdutoEdicaoEncalhe(FiltroConsultaEncalheDTO filtro, Long idTipoMovimento, boolean indQtdEncalheAposPrimeiroDia);
